+++ CopilotKit/examples/node/CHANGELOG.md
@@ -1,5 +1,24 @@
 # node
 
+## 1.16.0-mme-fix-tools-error.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/backend@0.4.0-mme-fix-tools-error.2
+
+## 1.16.0-beta-automation.1
+
+### Minor Changes
+
+- Test automation
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/backend@0.4.0-beta-automation.1
+  - @copilotkit/shared@0.4.0-beta-automation.1
+
 ## 1.16.0-tools.0
 
 ### Minor Changes